adding proper summaries, even for multi track activities

+++ com.github.rasifix.trainings.format.tcx/src/main/java/com/github/rasifix/trainings/format/tcx/TcxActivityReader.java
@@ -138,10 +138,10 @@ public class TcxActivityReader implements ActivityReader {
 		}
 		
 		if (source.getBikeCadence() != null) {
-			result.addAttribute(new CadenceAttribute((double) source.getBikeCadence()));
+			result.addAttribute(new CadenceAttribute(source.getBikeCadence()));
 			
 		} else if (source.getRunCadence() != null) {
-			result.addAttribute(new CadenceAttribute((double) source.getRunCadence()));
+			result.addAttribute(new CadenceAttribute(source.getRunCadence()));
 		}
 		
 		if (source.getSpeed() != null) {
